+++ runtime/runtime-core/common/src/aws/smithy/kotlin/runtime/content/ByteStream.kt
@@ -4,10 +4,11 @@
  */
 package aws.smithy.kotlin.runtime.content
 
-import aws.smithy.kotlin.runtime.io.SdkByteReadChannel
-import aws.smithy.kotlin.runtime.io.SdkSource
-import aws.smithy.kotlin.runtime.io.readToBuffer
-import aws.smithy.kotlin.runtime.io.readToByteArray
+import aws.smithy.kotlin.runtime.io.*
+import aws.smithy.kotlin.runtime.io.internal.SdkDispatchers
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.launch
 
 /**
  * Represents an abstract read-only stream of bytes
@@ -104,5 +105,94 @@ public fun ByteStream.cancel() {
         is ByteStream.Buffer -> stream.bytes()
         is ByteStream.ChannelStream -> stream.readFrom().cancel(null)
         is ByteStream.SourceStream -> stream.readFrom().close()
+    }
+}
+
+/**
+ * Return a [Flow] that consumes the underlying [ByteStream] when collected.
+ *
+ * @param bufferSize the size of the buffers to emit from the flow. All buffers emitted
+ * will be of this size except for the last one which may be less than the requested buffer size.
+ * This parameter has no effect for the [ByteStream.Buffer] variant. The emitted [ByteArray]
+ * will be whatever size the in-memory buffer already is in that case.
+ */
+public fun ByteStream.toFlow(bufferSize: Long = 8192): Flow<ByteArray> = when (this) {
+    is ByteStream.Buffer -> flowOf(bytes())
+    is ByteStream.ChannelStream -> readFrom().toFlow(bufferSize)
+    is ByteStream.SourceStream -> readFrom().toFlow(bufferSize).flowOn(SdkDispatchers.IO)
+}
+
+/**
+ * Create a [ByteStream] from a [Flow] of byte arrays.
+ *
+ * @param scope the [CoroutineScope] to use for launching a coroutine to do the collection in.
+ * @param contentLength the overall content length of the [Flow] (if known). If set this will be
+ * used as [ByteStream.contentLength]. Some APIs require a known `Content-Length` header and
+ * since the total size of the flow can't be calculated without collecting it callers should set this
+ * parameter appropriately in those cases.
+ */
+public fun Flow<ByteArray>.toByteStream(
+    scope: CoroutineScope,
+    contentLength: Long? = null,
+): ByteStream {
+    val ch = SdkByteChannel(true)
+    var totalWritten = 0L
+    val job = scope.launch {
+        collect { bytes ->
+            ch.write(bytes)
+            totalWritten += bytes.size
+
+            check(contentLength == null || totalWritten <= contentLength) {
+                "$totalWritten bytes collected from flow exceeds reported content length of $contentLength"
+            }
+        }
+
+        check(contentLength == null || totalWritten == contentLength) {
+            "expected $contentLength bytes collected from flow, got $totalWritten"
+        }
+
+        ch.close()
+    }
+
+    job.invokeOnCompletion { cause ->
+        ch.close(cause)
+    }
+
+    return object : ByteStream.ChannelStream() {
+        override val contentLength: Long? = contentLength
+        override val isOneShot: Boolean = true
+        override fun readFrom(): SdkByteReadChannel = ch
+    }
+}
+
+private fun SdkByteReadChannel.toFlow(bufferSize: Long): Flow<ByteArray> = flow {
+    val chan = this@toFlow
+    val sink = SdkBuffer()
+    while (!chan.isClosedForRead) {
+        val rc = chan.read(sink, bufferSize)
+        if (rc == -1L) break
+        if (sink.size >= bufferSize) {
+            val bytes = sink.readByteArray(bufferSize)
+            emit(bytes)
+        }
+    }
+    if (sink.size > 0L) {
+        emit(sink.readByteArray())
+    }
+}
+
+private fun SdkSource.toFlow(bufferSize: Long): Flow<ByteArray> = flow {
+    val source = this@toFlow
+    val sink = SdkBuffer()
+    while (true) {
+        val rc = source.read(sink, bufferSize)
+        if (rc == -1L) break
+        if (sink.size >= bufferSize) {
+            val bytes = sink.readByteArray(bufferSize)
+            emit(bytes)
+        }
+    }
+    if (sink.size > 0L) {
+        emit(sink.readByteArray())
     }
 }